Unit test for restaurant handler

+++ Application/CQRS/Restaurant/Handler/Command/DeleteRestaurantCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Application.CQRS.Restaurant.Request.Command;
+using Application.Exceptions;
 using Application.Persistence.Contracts;
 using AutoMapper;
 using MediatR;
@@ -18,6 +19,9 @@ namespace Application.CQRS.Restaurant.Handler.Command
         public async Task<Unit> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
             var restaurant = await _restaurantRepository.GetById(request.Id);
+            if (restaurant == null)
+                throw new NotFoundException(nameof(Domain.Entity.Restaurant), request.Id);
+
             await _restaurantRepository.Delete(restaurant);
             return Unit.Value;
         }
